add first version of timeline diagram
