+++ package.json
@@ -48,6 +48,7 @@
     "csv": "^6.1.0",
     "d3": "^7.4.4",
     "glob-promise": "^4.2.2",
+    "lit": "^2.3.1",
     "multistream": "^4.1.0"
   },
   "devDependencies": {
