+++ src/assets/js/charts/index.js
@@ -1,0 +1,2 @@
+export { default as generateTimelineChart } from './timeline';
+export { default as generateChordChart } from './chord';
