+++ src/views/index.html
@@ -37,7 +37,7 @@
         <input type="range" class="navigation-range" data-js-slider min="1">
       </div>
 
-      <timeline-player min="2019-01-01" value="2019-01-01" max="2021-12-28" step="w"></timeline-player>
+      <!-- <timeline-player min="2019-01-01" value="2019-01-01" max="2021-12-28" step="w"></timeline-player> -->
     </div>
 
   </main>
